+++ d3output.py
